rename Evaluators to avoid ambiguity
partly adresses #1124 (the Image folder still need a renames)
