+++ mteb/evaluation/evaluators/__init__.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from .AnySTSEvaluator import AnySTSEvaluator
-from .BitextMiningEvaluator import BitextMiningEvaluator
-from .ClassificationEvaluator import ClassificationEvaluator
-from .ClusteringEvaluator import ClusteringEvaluator
+from .sts_evaluator import AnySTSEvaluator
+from .bitextmining_evaluator import BitextMiningEvaluator
+from .classification_evaluator import ClassificationEvaluator
+from .clustering_evaluator import ClusteringEvaluator
 from .Evaluator import Evaluator
 from .Image import (
     Any2AnyMultiChoiceEvaluator,
@@ -11,27 +11,27 @@ from .Image import (
     ImageTextPairClassificationEvaluator,
 )
 from .dense_retrieval_exact_search import DenseRetrievalExactSearch
-from .PairClassificationEvaluator import PairClassificationEvaluator
-from .RetrievalEvaluator import RetrievalEvaluator
-from .SummarizationEvaluator import (
+from .pairclassification_evaluator import PairClassificationEvaluator
+from .retrieval_evaluator import RetrievalEvaluator
+from .summarization_evaluator import (
     DeprecatedSummarizationEvaluator,
     SummarizationEvaluator,
 )
-from .ZeroShotClassificationEvaluator import ZeroShotClassificationEvaluator
+from .zeroshot_classification_evaluator import ZeroShotClassificationEvaluator
 
 __all__ = [
     "Evaluator",
-    "AnySTSEvaluator",
-    "SummarizationEvaluator",
+    "sts_evaluator",
+    "summarization_evaluator",
     "DeprecatedSummarizationEvaluator",
-    "RetrievalEvaluator",
+    "retrieval_evaluator",
     "DenseRetrievalExactSearch",
-    "ClusteringEvaluator",
-    "BitextMiningEvaluator",
-    "PairClassificationEvaluator",
+    "clustering_evaluator",
+    "bitextmining_evaluator",
+    "pairclassification_evaluator",
     "Any2AnyMultiChoiceEvaluator",
     "Any2AnyRetrievalEvaluator",
     "ImageTextPairClassificationEvaluator",
-    "ZeroShotClassificationEvaluator",
-    "ClassificationEvaluator",
+    "zeroshot_classification_evaluator",
+    "classification_evaluator",
 ]
