+++ mteb/evaluation/evaluators/sts_evaluator.py
